==> notebooks/2_Create_Dataset3.ipynb <==
import polars as pl

aml_df = pl.read_parquet('../data/jp_morgan/transaction_metrics_final_aml_v3.parquet')
frd_df = pl.read_parquet('../data/jp_morgan/transaction_metrics_final_fraud_v3.parquet')

aml_df = aml_df.with_columns(
    pl.when(pl.col("Label") == "GOOD").then(0).otherwise(1).alias("Label")
)

aml_df = aml_df.with_columns(pl.col('Label').cast(pl.Int32)).sort("Time_step")
frd_df = frd_df.with_columns(pl.col('Label').cast(pl.Int32)).sort("Time_step")
ds3_df = pl.concat([aml_df, frd_df]).sort("Time_step")

ds3_df.write_parquet('../data/jp_morgan/sorted/transaction_metrics_final_ds3_v3.parquet')

ds3_df.describe()
