==> notebooks/3_preprocess_Graph_Extract_Edges.ipynb <==
import pandas as pd

data_dir = '../data/jp_morgan'
data_ = 'aml'

sampling = False
if sampling:
    data_name = f'sample_{data_}'
else:
    data_name = f'final_{data_}'

df = pd.read_parquet(f'{data_dir}/sorted/transaction_metrics_{data_name}.parquet')

test_revert = df.copy()

df.columns

def create_account_id(id_col, account_col):
    if pd.isna(id_col) and pd.isna(account_col):
        return None
    elif pd.isna(id_col):
        return account_col
    elif pd.isna(account_col):
        return id_col
    else:
        return f'{id_col}_{account_col}'

df['party_account_id'] = df.apply(lambda row: create_account_id(row['party_Id'], row['party_Account']), axis=1)
df['cparty_account_id'] = df.apply(lambda row: create_account_id(row['cparty_Id'], row['cparty_Account']), axis=1)


output_df = df[[
    'party_account_id',
    'cparty_account_id',
    'Transaction_Id',
    'USD_amount',
    'txn_time_hr',
    'txn_time_mm',
    'txn_age_days',
    'is_credit',
    'std_txn_type'
    ]]

output_df.columns = ['party_id', 'counterparty_id', 'transaction_id', 'amount',
                     'txn_time_hr', 'txn_time_mm', 'txn_age_days', 'is_credit',
                     'std_txn_type']

output_df.to_parquet(f'{data_dir}/sorted/edges_{data_name}.parquet')

def extract_entities(df):
    df1 = df.groupby('party_Id').agg({'party_Country': lambda x: pd.Series.mode(x)[0] if not pd.Series.mode(x).empty else None})
    party_df = df[['party_Id', 'party_Account']].join(df1, on='party_Id')
    party_df.columns = ['Id', 'Account', 'Country']

    df2 = df.groupby('cparty_Id').agg({'cparty_Country': lambda x: pd.Series.mode(x)[0] if not pd.Series.mode(x).empty else None})
    cparty_df = df[['cparty_Id', 'cparty_Account']].join(df2, on='cparty_Id')
    cparty_df.columns = ['Id', 'Account', 'Country']

    return pd.concat([party_df, cparty_df]).drop_duplicates()

entities = extract_entities(df)
entities.to_parquet(f'{data_dir}/sorted/entities_{data_name}.parquet')

def extract_entities(df):
    df1 = df.groupby('party_Id').agg({'party_Country': lambda x: pd.Series.mode(x)[0] if not pd.Series.mode(x).empty else None})
    party_df = df[['party_Id', 'party_Account']].join(df1, on='party_Id')
    party_df.columns = ['Id', 'Account', 'Country']

    df2 = df.groupby('cparty_Id').agg({'cparty_Country': lambda x: pd.Series.mode(x)[0] if not pd.Series.mode(x).empty else None})
    cparty_df = df[['cparty_Id', 'cparty_Account']].join(df2, on='cparty_Id')
    cparty_df.columns = ['Id', 'Account', 'Country']

    return pd.concat([party_df, cparty_df]).drop_duplicates()

entities = extract_entities(sample)
entities.to_parquet(f'{data_dir}/sorted/entities_sample_{data_name}.parquet')
